move eror notification to molecules notifications
