+++ src/pages/Davit.tsx
@@ -4,7 +4,7 @@ import { useDispatch } from "react-redux";
 // import {BrowserRouter, Route, Switch} from "react-router-dom";
 // electron needs HashRouter
 import { HashRouter as BrowserRouter, Route, Switch } from "react-router-dom";
-import { ErrorNotification } from "../components/common/fragments/ErrorNotification";
+import { ErrorNotification } from "../components/molecules/notification/ErrorNotification";
 import { ControlPanelController } from "../components/organisms/controllPanel/presentation/ControlPanelController";
 import { SidePanelController } from "../components/organisms/sidePanel/SidePanelController";
 import "../css/Davit.css";
